pyproject.toml
[project]
name = "dashboard-ki-agenten"
version = "0.1.0"
description = "DBS Home Dashboard fuer KI-gestuetzte Verwaltungs-Workflows"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.17",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "alembic>=1.14",
    "authlib>=1.3",
    "itsdangerous>=2.2",
    "httpx>=0.27",
    "anthropic>=0.40",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "schwifty>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
